databinding으로 선택 회차 로또 추첨 데이터 뿌리기

+++ app/src/main/java/com/zahi/lotto/api/LottoApi.kt
@@ -11,7 +11,7 @@ interface LottoApi {
     @GET("common.do")
     fun getLottoWinnerNumber(
         @Query("method") method: String = "getLottoNumber",
-        @Query("drwNo") drwNo: Int,
+        @Query("drwNo") drwNo: Long,
     ): Observable<LotteryNumber>
 
 }
